Cloned Dec 18 version of SandBox script
Developed the original version of the skeena script further for integration in the DFO SandBox. This extrapolated the analysis further making it easier to script and save multiple runs.

Relative file structured used in the SandBox has been retained.

Added data so that all dependencies are cloned/copied into this repository.